+++ src/utils/index.ts
@@ -1,5 +1,1 @@
-export function objectToCssString(obj: Record<string, any>): string {
-  return Object.entries(obj)
-    .map(([key, value]) => `${key}: ${value};`)
-    .join(' ');
-}
+export * from './object-to-css';
