+++ src/utils/object-to-css.ts
@@ -1,0 +1,5 @@
+export function objectToCssString(obj: Record<string, any>): string {
+  return Object.entries(obj)
+    .map(([key, value]) => `${key}: ${value};`)
+    .join(' ');
+}
